# docs/swagger.yaml
basePath: /
definitions:
...
  handler.HTTPError:
    properties:
      error:
        type: string
    type: object
  handler.HTTPMessage:
    properties:
      message:
        type: string
    type: object
...
  handler.ListRequest:
    properties:
      filter:
        additionalProperties:
          type: string
        type: object
      length:
        type: integer
      sort:
        additionalProperties:
          type: string
        type: object
      start:
        type: integer
    type: object
...
  handler.NewOption:
    properties:
      description:
        type: string
      name:
        type: string
      sort:
        type: integer
      standard:
        type: boolean
      title:
        type: string
      valueId:
        type: integer
    type: object
...
  handler.OptionFullView:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      sort:
        type: integer
      standard:
        type: boolean
      title:
        type: string
      valueId:
        type: integer
      values:
        items:
          $ref: '#/definitions/handler.ValueView'
        type: array
    type: object
...
  handler.OptionShortView:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      sort:
        type: integer
      standard:
        type: boolean
      title:
        type: string
      valueId:
        type: integer
    type: object
  handler.OptionView:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      sort:
        type: integer
      standard:
        type: boolean
      title:
        type: string
      value:
        $ref: '#/definitions/handler.ValueView'
        type: object
      values:
        items:
          $ref: '#/definitions/handler.ValueView'
        type: array
    type: object
  handler.OptionsFullView:
    items:
      $ref: '#/definitions/handler.OptionFullView'
    type: array
  handler.OptionsListItem:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      sort:
        type: integer
      standard:
        type: boolean
      title:
        type: string
      valueValue:
        type: string
      valuesValues:
        type: string
    type: object
  handler.OptionsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handler.OptionsListItem'
        type: array
      filtered:
        type: integer
      total:
        type: integer
    type: object
...
  handler.ValueView:
    properties:
      availability:
        type: string
      description:
        type: string
      id:
        type: integer
      sending:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      value:
        type: string
    type: object
...
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: GoShop full featured api, see documentation https://docs.google.com/document/d/1VlkAYTqZG9oGvZxnSNDpcuafU-msOb2Qm2FevMJkA7w/edit
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GoShop API
  version: "1.0"
paths:
...
  /api/v1/options:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OptionsFullView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Get options
      tags:
      - option
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: option
        required: true
        schema:
          $ref: '#/definitions/handler.NewOption'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OptionView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Create option
      tags:
      - option
  /api/v1/options/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Option ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.HTTPMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Delete option
      tags:
      - option
    get:
      consumes:
      - application/json
      parameters:
      - description: Option ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OptionView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Get option
      tags:
      - option
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: option
        required: true
        schema:
          $ref: '#/definitions/handler.OptionShortView'
      - description: Option ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OptionShortView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Update option
      tags:
      - option
  /api/v1/options/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OptionsListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.HTTPError'
      security:
      - BasicAuth: []
      summary: Search options
      tags:
      - option
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"